+++ CHANGELOG.md
@@ -16,6 +16,16 @@ All notable changes to this project will be documented in this file.
   N-fermion subspace of a full Hilbert space. In addition to the view classes,
   there are two new utility functions ``loperator.n_fermion_sector_size()`` and
   ``loperator.n_fermion_sector_basis_states()``.
+- New classes ``loperator.NFermionMultiSectorView(R|C)``. These objects are
+  ``LOperator(R|C)``-compatible views that represent state vectors defined in an
+  N-fermion multisector. A multisector is a set of all basis states, which have
+  ``N_1`` particles within a subset of fermionic modes ``{S_1}``, ``N_2``
+  particles within another subset ``{S_2}`` and so on. There can be any number
+  of individual pairs ``({S_i}, N_i)`` (sectors contributing to the multisector)
+  as long as all subsets ``{S_i}`` are disjoint. In addition to the view
+  classes, there are two new utility functions
+  ``loperator.n_fermion_multisector_size()`` and
+  ``loperator.n_fermion_multisector_basis_states()``.
 - New method ``loperator.HilbertSpace.has_algebra()``.
 - New method ``loperator.SpacePartition.find_connections()``.